// src/api/enums/CommentCategory.ts
// Copyright (c) 2017-2020, The Particl Market developers
// Distributed under the GPL software license, see the accompanying
// file COPYING or https://github.com/particl/particl-market/blob/develop/LICENSE

/**
 * CommentType
 *
 */

export enum CommentCategory {
    LISTINGITEM_QUESTION_AND_ANSWERS = 'LISTINGITEM_QUESTION_AND_ANSWERS',
    PROPOSAL_QUESTION_AND_ANSWERS = 'PROPOSAL_QUESTION_AND_ANSWERS',
    MARKETPLACE_COMMENT = 'MARKETPLACE_COMMENT',
    PRIVATE_MESSAGE = 'PRIVATE_MESSAGE'
}